Implementado validar!

+++ ruby/lib/Person.rb
@@ -5,7 +5,7 @@ class Person
 
   has_one String, named: :first_name
   has_one String, named: :last_name
-  has_one Numeric, named: :age
+  has_one Integer, named: :age
   has_one Boolean, named: :is_admin
 
   attr_accessor :first_name, :last_name, :age, :is_admin
